Replace the ability to pass a plain string and get a plain TEXT email.
This is to allow people to call zen_mail_from_template() as they did with
zen_mail(), passing a simple string instead of a data block, and get a
plain text email.  Mostly for admin emails that do not need wrapping in a
pretty header/footer email with text/html parts.
This bypasses all template processing and avoids the HTML body part entirely.