Provide CONTENT_* notifiers and improve notifier efficiency.
Add NotifierManager::insertContent() to provide for hooks in pages to
produce output either from define pages, or from observers that take action
based on data (possibly resulting in outputting define pages).
Maintain a separate array of hashes and eventID => observers, so that when
we need to look up observers for an eventID it's a simple array lookup.